chore: factored out Profile and PostSurvey types
